# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FleetVision AI",
        "description": "FleetVision AI Backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/alerts": {
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Create Alert",
                "operationId": "create_alert_api_v1_alerts_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlertCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get Alerts",
                "operationId": "get_alerts_api_v1_alerts_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/status/active": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get Active Alerts",
                "operationId": "get_active_alerts_api_v1_alerts_status_active_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResponse"
                                    },
                                    "title": "Response Get Active Alerts Api V1 Alerts Status Active Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/vehicle/{vehicle_id}/active": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get Active Vehicle Alerts",
                "operationId": "get_active_vehicle_alerts_api_v1_alerts_vehicle__vehicle_id__active_get",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Vehicle Id"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResponse"
                                    },
                                    "title": "Response Get Active Vehicle Alerts Api V1 Alerts Vehicle  Vehicle Id  Active Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/vehicle/{vehicle_id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get Vehicle Alerts",
                "operationId": "get_vehicle_alerts_api_v1_alerts_vehicle__vehicle_id__get",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Vehicle Id"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResponse"
                                    },
                                    "title": "Response Get Vehicle Alerts Api V1 Alerts Vehicle  Vehicle Id  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/filter/status/{alert_status}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get Alerts By Status",
                "operationId": "get_alerts_by_status_api_v1_alerts_filter_status__alert_status__get",
                "parameters": [
                    {
                        "name": "alert_status",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/AlertStatus"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResponse"
                                    },
                                    "title": "Response Get Alerts By Status Api V1 Alerts Filter Status  Alert Status  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/filter/severity/{severity}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get Alerts By Severity",
                "operationId": "get_alerts_by_severity_api_v1_alerts_filter_severity__severity__get",
                "parameters": [
                    {
                        "name": "severity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/AlertSeverity"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResponse"
                                    },
                                    "title": "Response Get Alerts By Severity Api V1 Alerts Filter Severity  Severity  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/filter/type/{alert_type}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get Alerts By Type",
                "operationId": "get_alerts_by_type_api_v1_alerts_filter_type__alert_type__get",
                "parameters": [
                    {
                        "name": "alert_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/AlertType"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResponse"
                                    },
                                    "title": "Response Get Alerts By Type Api V1 Alerts Filter Type  Alert Type  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/alerts/{alert_id}/acknowledge": {
            "patch": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Acknowledge Alert",
                "operationId": "acknowledge_alert_api_v1_alerts__alert_id__acknowledge_patch",
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Alert Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/{alert_id}/resolve": {
            "patch": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Resolve Alert",
                "operationId": "resolve_alert_api_v1_alerts__alert_id__resolve_patch",
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Alert Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/{alert_id}/status": {
            "patch": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Update Alert Status",
                "operationId": "update_alert_status_api_v1_alerts__alert_id__status_patch",
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Alert Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlertStatusUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AlertCreate": {
                "properties": {
                    "vehicle_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Vehicle Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "alert_type": {
                        "$ref": "#/components/schemas/AlertType"
                    },
                    "severity": {
                        "$ref": "#/components/schemas/AlertSeverity"
                    }
                },
                "type": "object",
                "required": [
                    "vehicle_id",
                    "title",
                    "description",
                    "alert_type",
                    "severity"
                ],
                "title": "AlertCreate"
            },
...
            "AlertSeverity": {
                "type": "string",
                "enum": [
                    "LOW",
                    "MEDIUM",
                    "HIGH",
                    "CRITICAL"
                ],
                "title": "AlertSeverity"
            },
...
            "AlertStatus": {
                "type": "string",
                "enum": [
                    "ACTIVE",
                    "ACKNOWLEDGED",
                    "RESOLVED"
                ],
                "title": "AlertStatus"
            },
...
            "AlertType": {
                "type": "string",
                "enum": [
                    "SPEEDING",
                    "FATIGUE",
                    "COLLISION",
                    "MAINTENANCE",
                    "ENGINE",
                    "FUEL",
                    "BATTERY",
                    "GEOFENCE",
                    "AI_RISK"
                ],
                "title": "AlertType"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    },
...
}